requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
joblib==1.3.2
scikit-learn==1.4.0
